Move realsense functionality from camera node to realsense module
